Move getRoom(s) methods to RoomController
